refactor(rpc): merge rpc-api + rpc-transport into single rpc.ts
Two files (192 LOC total) for a Cap'n Web RPC feature that's half-wired:
DodoPublicApi was exported but never instantiated in production src/;
AgentConnectionTransport fed messages to a phantom RpcSession that was
never constructed. The split between 'api' and 'transport' was scaffolding
for a future feature that hasn't shipped — but the two-file pretence
implied more structure than existed.

Merge into src/rpc.ts with a header comment marking it as scaffolding.
Keep DodoPublicApi because it IS exercised in tests
(security-audit-fixes-unit.test.ts and realtime.test.ts).
Keep DodoAuthenticatedApi (used by /rpc route) and AgentConnectionTransport
(instantiated by coding-agent.ts WS branch — current no-op behaviour
preserved).

Per the improve-codebase-architecture audit, candidate #4.
